pyproject.toml
[project]
name = "trm"
version = "0.1.0"
description = "Adaptive Recursive CNN for Facial Expression Recognition"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "albumentations>=1.3.0",
    "timm>=0.9.0",
    "wandb>=0.15.0",
    "scikit-learn>=1.3.0",
    "pandas>=1.5.0",
    "numpy>=1.24.0",
    "pillow>=9.5.0",
    "tqdm>=4.64.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
]
